+++ src/web/wizard/components/Description.jsx
@@ -18,9 +18,8 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import createReactClass from 'create-react-class';
-import {Input} from 'components/bootstrap';
-import { Row, Col } from 'components/graylog';
-import {FormattedMessage} from 'react-intl';
+import { Input, Row, Col } from 'components/bootstrap';
+import { FormattedMessage } from 'react-intl';
 
 const Description = createReactClass({
     displayName: 'Description',