+++ src/web/wizard/components/TitleSeverity.jsx
@@ -18,10 +18,9 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import createReactClass from 'create-react-class';
-import { Input } from 'components/bootstrap';
+import { Input, Row, Col } from 'components/bootstrap';
 import { Select } from 'components/common';
-import { Row, Col } from 'components/graylog';
-import {FormattedMessage} from 'react-intl';
+import { FormattedMessage } from 'react-intl';
 
 const TitleSeverity = createReactClass({
     displayName: 'TitleSeverity',