+++ src/web/wizard/components/conditions/StatisticalCondition.jsx
@@ -19,8 +19,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import createReactClass from 'create-react-class';
 import { FormattedMessage } from 'react-intl';
-import { Row, Col } from 'components/graylog';
-import { Input} from 'components/bootstrap';
+import { Input, Row, Col } from 'components/bootstrap';
 import { Select } from 'components/common';
 import ObjectUtils from 'util/ObjectUtils';