+++ src/web/wizard/components/ruletype/OrCondition.jsx
@@ -21,13 +21,13 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import createReactClass from 'create-react-class';
 import ObjectUtils from 'util/ObjectUtils';
-import {FormattedMessage} from 'react-intl';
+import { FormattedMessage } from 'react-intl';
 import TitleSeverity from 'wizard/components/TitleSeverity';
 import FieldsCondition from 'wizard/components/conditions/FieldsCondition';
 import NumberCondition from 'wizard/components/conditions/NumberCondition';
 import TimeRangeCondition from 'wizard/components/conditions/TimeRangeCondition';
 import Description from 'wizard/components/Description';
-import { Row, Col } from 'components/graylog';
+import { Row, Col } from 'components/bootstrap';
 
 const STREAM = {
         matching_type: '',