+++ src/web/wizard/pages/ImportListPage.jsx
@@ -20,7 +20,7 @@ import createReactClass from 'create-react-class';
 import Routes from 'routing/Routes';
 import { addLocaleData, IntlProvider, FormattedMessage } from 'react-intl';
 import { LinkContainer } from 'react-router-bootstrap';
-import { Row, Col, Button } from 'components/graylog';
+import { Row, Col, Button } from 'components/bootstrap';
 import messages_fr from 'translations/fr.json';
 import AlertListActions from 'wizard/actions/AlertListActions';
 import { DocumentTitle, PageHeader } from 'components/common';