+++ src/web/wizard/pages/UpdateAlertPage.jsx
@@ -18,12 +18,12 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import createReactClass from 'create-react-class';
-import {Button, Col, Row} from 'components/graylog';
-import {LinkContainer} from 'react-router-bootstrap';
-import {DocumentTitle, PageHeader, Spinner} from 'components/common';
+import { Button, Col, Row } from 'components/bootstrap';
+import { LinkContainer } from 'react-router-bootstrap';
+import { DocumentTitle, PageHeader, Spinner } from 'components/common';
 import AlertRuleForm from 'wizard/components/AlertRuleForm';
 import AlertRuleActions from 'wizard/actions/AlertRuleActions';
-import {addLocaleData, IntlProvider, FormattedMessage} from 'react-intl';
+import { addLocaleData, IntlProvider, FormattedMessage } from 'react-intl';
 import messages_fr from 'translations/fr.json';
 import withParams from 'routing/withParams';
 import Navigation from 'wizard/routing/Navigation';